New file: pathway scoring
In this code we add to the original database some features related with the expression of the metabolic pathways listed in the variable "paths". The file where it come from is obtained from an extensive revision in genomic references. I score a path additively in the following way: if an indiviudual have a SNP from the list of a pathway with at least one mutant allele I increase the score by one.